requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.4
langchain==0.3.7
langchain_community==0.3.5
langchain_core==0.3.15
langchain_openai==0.2.6
langchain_tools==0.1.34
langgraph==0.2.45
langserve==0.3.0
Markdown==3.6
matplotlib==3.9.2
pdfkit==1.0.0
pydantic==2.9.2
python-dotenv==1.0.1
python_docx==1.1.2
Requests==2.32.3
uvicorn==0.32.0
